chore(deps): Include recommended lint configs
For now, this config won't use any custom rules. Instead, this will
serve as a single entry point for combining several existing
configurations, namely ESLint's own recommended configs, JavaScript
Standard Style, and Vue.js recommended configs.

The point of this config is to allow me a single repository where
linting configurations get updated, and not necessarily to have my own
custom style. The included configs work for me just fine. Another
benefit of this repository is the ability to hide the plugins being used
and only require ESLint itself as a peer dependency. The plugins are
purposefully listed as dependencies so they are automatically installed
and packages using this config do not need to worry about installing
them as well.