chore(package): Test on `osx`
Though in the case for this articular package, testing on multiple
platforms may be unnecessary, I still like to make a point to test on
all platforms where a package can be used. Therefore, tests for this
package should run on all available platform types (Travis CI still does
not currently support Windows).